+++ ProgramadorGenetico/src/es/si/ProgramadorGenetico/Interfaz/Pruebas/PruebaDibujanteNuevo1.java
@@ -16,9 +16,9 @@ import javax.swing.JPanel;
 
 public class PruebaDibujanteNuevo1 {
 
-	private static double[][][] transiciones;
+	private static float[][][] transiciones;
 	
-	private static double[] probabilidadFinal;
+	private static float[] probabilidadFinal;
 	
 	private static int estados;
 		
@@ -39,14 +39,14 @@ public class PruebaDibujanteNuevo1 {
 	public static void setValoresEntrada() {
 		
 		estados = 4;
-		transiciones= new double[estados][2][estados+1];
-		transiciones[0][1][0]=1.0;
-		transiciones[1][1][0]=0.4;
-		transiciones[2][1][0]=1.0;
-		transiciones[2][0][0]=1.0;
-		transiciones[3][1][0]=1.0;
-		transiciones[1][0][2]=0.9;
-		transiciones[2][1][2]=0.3;		
+		transiciones= new float[estados][2][estados+1];
+		transiciones[0][1][0]=1.0f;
+		transiciones[1][1][0]=0.4f;
+		transiciones[2][1][0]=1.0f;
+		transiciones[2][0][0]=1.0f;
+		transiciones[3][1][0]=1.0f;
+		transiciones[1][0][2]=0.9f;
+		transiciones[2][1][2]=0.3f;		
 			
 	}
